Move validation tests to store package
